+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.0 (November 8, 2017)
+
+#### 🚀 New Feature
+
+- `app`
+  - Add WordPress type choice: single-site or multisite. ([@ruslankhh](https://github.com/ruslankhh))
+  - Change WordPress file structure (`app` folder now includes `core`, `plugins` and `themes` folders). ([@ruslankhh](https://github.com/ruslankhh))
+
+#### Committers: 1
+
+- Ruslan Khusnetdinov ([ruslankhh](https://github.com/ruslankhh))
+
 ## 0.2.0 (November 6, 2017)
 
 #### 🚀 New Feature